Finish landing-pg exercise
Wow! What a tough project. It wasn't too tough to create the original layout for the web, but once I thought about how to make it responsive, my code started to break. As a result, I had to nest divs and more wrappers than I thought was necessary. The website is still rough around the edges and could use a lot more polish, but it's time for me to  move on! 😤